[core] remove legacy start/stop/status commands
Every component is now handled by `supervisor`, it's time to remove
these commands.

They were deprecated with 5.4.0, more than 1 year ago.
